1. 高程值统一使用float类型 2. 山体阴影实现迁移到silly_dem中

+++ su/metadata/silly_ascii_grid.cpp
@@ -237,7 +237,7 @@ bool suAsciiGrid::write_asc(const suPath& file) const
     std::stringstream ssGrid;
     std::string fillV = SUFMT("{:.3f}", info.fill);
     ssGrid << std::fixed << std::setprecision(3);
-    double* p = raster.data();
+    float* p = raster.data();
     for (int r = 0; r < info.height; ++r)
     {
         for (int c = 0; c < info.width; ++c)
